Move all tile sources to function sources (#634)
This keeps the tile sources consistent (currently it is a mix of views
and functions), and allows using the 'id' column of a layer as ID,
without promoting it client-side.

Changing the fields of a single tile view is also possible by redefining
the tile function (and comment), without further changing the Martin
configuration or even restarting the Martin container.