Yards: use total yard rail length as importance (#679)
Part of https://github.com/hiddewie/OpenRailwayMap-vector/issues/662

Analysis in
https://github.com/hiddewie/OpenRailwayMap-vector/issues/662#issuecomment-3566983532

This pull request uses the total rail length of all railway lines within
a yard, to assign a measure of importance to the yard.

This is similar to how platforms and railway routes assign importance to
passenger railway stations.

The current thresholds are set at >= 7200 meters of rail is a large
yard, >= 750 meters of rail is a normal yard, and all others are small.

I renamed everything *route count* to *importance*, because stations and
yards each get a (scaled) importance which can be compared (e.g. in API
search results).