Known operators (#707)
For #597 

Goal: add a fixed set of operators with a predefined list of metadata,
stored in
https://github.com/hiddewie/OpenRailwayMap-vector/blob/known-operators/features/operator.yaml.

Not every operator (thousands in the world) needs to be defined. Those
which are not defined will still get the same automatic generated color
like before.

Additionally, multiple operators on stations, and multiple networks on
stations are supported.

---------

Co-authored-by: wolfy1339 <4595477+wolfy1339@users.noreply.github.com>