Add colours for some German operators (#717)
This adds colours for some German operators.

The important part is DB InfraGO, the main operator of the German rail
network. Since this company is only two years old in its current form,
there are several different values in use for the operator tag which
currently are rendered in different colours. This should be fixed.

I used this opportunity to also assign colours to some other companies
where the current colours don't fit or clash with similar colours of
different operators in the same area. I mainly used the colours from the
companies' logos. Where this conflicts with other colours of operators
active in the same region (in particular DB's red), I used other colours
associated with the company (e.g. based on the vehicle livery in
Karlsruhe instead of the red logo).

**List of colours:**

_Railway:_
- DB InfraGO `#f01414`
- DB RegioNetz `#dd4488`
- AVG `#008833`
- Regio Infra Nord-Ost `#007733`
- SWEG `#114499`
- UBB `#114488`

_Tram/subway:_
- Berlin `#f0d020`
- Frankfurt `#008080`
- Karlsruhe `#ffaa33`
- Rhein-Neckar `#ee7700`
- Rostock `#005090`
- Stuttgart `#ffcc00`

Co-authored-by: Hidde Wieringa <hidde@hiddewieringa.nl>